+++ NEWS.md
@@ -1,14 +1,29 @@
-# development version
+# tidywater development version
+
+* `calculate_corrosion` expanded CCPP search range (fewer errors)
+
+## Breaking changes
+* `pac_toc` now constrained to accept a smaller range of doses and times. The form of the equation was allowing negative TOC outputs.
+* `define_water`, `chemdose_ph`, `blend_waters`, and other pH related functions slightly impacted by a fix in concentration vs activity. 
+Previous code assumed pH = 10^-[H+], code has been corrected to pH = 10^-{H+}
+
+## New features
+* User may now choose which cation or anion to use for balancing ions
+
+# tidywater 0.7.0
 
 ## New features
 * chlorine and chloramine decay: `chemdose_chlordecay`
 * New water slots for chloramine chemistry: `combined_chlorine`, `nh2cl`, `nhcl2`, `ncl3`
+* `solvemass_solids` separates functionality from `solvecost_solids` to solve lb/day
+* `biofilter_toc`, `chemdose_chlordecay`, `ozonate_bromate`, and `solvect` helpers now available.
 
 ## Breaking changes
 * `chemdose_ct` renamed `solvect_chlorine`
 * `ozonate_ct` renamed `solvect_o3`
 * `tot_ocl` slot in water renamed `free_chlorine`
 * `define_water` argument changes: `tot_ocl` changed to `free_chlorine`, added `combined_chlorine`
+* Helper function (`_chain` and `_once`) behavior change: if multiple values are specified for multiple arguments, all combinations are used.
 
 # tidywater 0.6.2